Some parsing cleanup

+++ tower-http/src/services/fs/mod.rs
@@ -146,7 +146,7 @@ where
         max_read_bytes: u64,
     ) -> AsyncReadBody<Take<T>> {
         AsyncReadBody {
-            reader: ReaderStream::with_capacity(read.take(max_read_bytes as u64), capacity),
+            reader: ReaderStream::with_capacity(read.take(max_read_bytes), capacity),
         }
     }
 }
